+++ src/model/error_proto.rs
@@ -1,3 +1,4 @@
+use std::fmt::{Display, Formatter};
 use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Default, Clone, Serialize, Deserialize)]
@@ -15,4 +16,37 @@ pub struct ErrorProto {
     /// A short error code that summarizes the error.
     #[serde(skip_serializing_if = "Option::is_none")]
     pub reason: Option<String>,
+}
+
+impl Display for ErrorProto {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        let mut buffer = String::new();
+
+        if let Some(debug_info) = &self.debug_info {
+            buffer += &format!("debug_info: {}", debug_info);
+        }
+
+        if let Some(location) = &self.location {
+            if buffer.len() > 0 {
+                buffer += ", ";
+            }
+            buffer += &format!("location: {}", location);
+        }
+
+        if let Some(message) = &self.message {
+            if buffer.len() > 0 {
+                buffer += ", ";
+            }
+            buffer += &format!("message: {}", message);
+        }
+
+        if let Some(reason) = &self.reason {
+            if buffer.len() > 0 {
+                buffer += ", ";
+            }
+            buffer += &format!("reason: {}", reason);
+        }
+
+        f.write_str(&format!("ErrorProto: {{{}}}", buffer))
+    }
 }