+++ src/model/table_data_insert_all_request.rs
@@ -60,4 +60,16 @@ impl TableDataInsertAllRequest {
         self.template_suffix = Some(suffix.into());
         self
     }
+
+    pub fn is_empty(&self) -> bool {
+        self.rows.is_empty()
+    }
+
+    pub fn len(&self) -> usize {
+        self.rows.len()
+    }
+
+    pub fn clear(&mut self) {
+        self.rows.clear()
+    }
 }