+++ CHANGELOG.md
@@ -6,11 +6,13 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
+- Add methods `is_empty`, `len`, `clear` to `TableDataInsertAllRequest`
 - Add method `add_rows` to `TableDataInsertAllRequest` (thanks @nixxholas)
 - Bump version for yup-oauth2 v6 (thanks @JamesHinshelwood)
-- Implement the Default trait for most of the structures in the model sub-directory
-- Implement the Clone trait for most of the structures in the model sub-directory
-
+- Implement `Default` trait for most of the structures in the model sub-directory
+- Implement `Clone` trait for most of the structures in the model sub-directory
+- Implement `Display` trait for `ErrorProto` and `TableDataInsertAllResponseInsertErrors`
+- 
 ## [0.9.3] - 2021-08-31
 
 ### Fix